Initialize environment before loading library code
When `lib/language_pack/base.rb` is evaluated, it tries to define the
`VENDOR_URL` constant by looking in the environment. However in the case where
`user-env-compile` is not enabled, the environment won't yet have been copied
from the filesystem. So, require the `shell_helpers` and initialize the
environment before loading the rest of the library code.

Note this does mean that the environment will *always* be initialized,
regardless of whether the language pack matches, but this doesn't seem like it
would hurt anything.