Raise error if rake file exists but cannot load
In Rails 3+ it is a reasonable assumption that most developers want to run `assets:precompile` if the task is present. If they accidentally introduce an error into their Rakefile then we cannot detect properly we should raise instead of continuing. Behavior is the same for regular Ruby projects and Rails2.