Don't force the postgresql addons on Rails apps
This will revert to the global ruby default where the addon is inserted only
if the app has a 'pg' gem in it's Gemfile.

A rails app which uses an external database like RDS will get it's
config/database.yml overwritten because of the PG addon and this will break
the app. Combined with the fact that the new pipeline feature and the app.json
format has no way of disabling the default addons, it means that new review
apps always get broken by default.