[close #507] Show bundler warning earlier
We should be showing a bundler upgrade warning as soon as possible. Currently if you select a "bad" version of ruby in your Gemfile.lock you will not get this error. This is a problem for users who are running into https://devcenter.heroku.com/articles/bundler-version#known-upgrade-issues, while the bug is well known and documented, they're not being directed to that page. 

This patch will emit an upgrade warning for bundler before we even try downloading it. This should ensure the warning is always present in build outputs.