Don't rely on system library `json`
This is super UGHHH. There are cases when a `json` gem has been installed already. One case is when two ruby buildpacks run.

Here's what I think is happening: The first buildpack sets the PATH so that future buildpacks can execute Ruby code. This means that now all ruby gems installed from the customer app are on GEM_PATH. Now when the next invocation of building ruby comes up it fires `require 'json'` ruby sees that there's a gem installed and tries to load it, unfortunately it's for a different version of ruby than the one the buildpack is running on.

One option to solve is by not requiring `json` but to vendor in a json library such as https://github.com/kr/okjson.