Make the Ruby buildpack work with Ruby vendored Bundler.
With Bundler being vendored in Ruby, we have no choice but to use the
Ruby vendored Bundler if it's >= than the one we package in the Ruby
buildpack since Ruby comes first in the `GEM_PATH` and `require
'bundler'` will always pull the latest. This still allows us the ability
to update Bundler when we need to. Forcing the version of Bundler with
`bundle _1.15.2_` won't be propagated during runtime. Having a different
version used during build and runtime is not a good idea.

This patch works also works around the [binstub `Kernel#load`
issue](https://bugs.ruby-lang.org/issues/13997) in preview1.