Disable listing check for disabling cops
For a backstory see https://github.com/bbatsov/rubocop/issues/4459

You can configure rubocop to error out when anyone disables rubocop errors. Which is fine, except the Ruby buildpack needs to do that. We needed a way to disable that check, so this feature was added https://github.com/bbatsov/rubocop/pull/4938.

This PR is using that feature to hopefully allow anyone to deploy and not get errors from rubocop.