Fix Rails Runner timeout
IO.popen cannot be timed out reliably: https://stackoverflow.com/questions/17237743/timeout-within-a-popen-works-but-popen-inside-a-timeout-doesnt

The issue is that the process can do something like `Process.spawn && wait` which defers control to a different process. Ruby cannot wake the process up in order to terminate it so when `io.close` is called (at the end of the block), the process deadlocks on the child process being terminated which may never happen.

The `run` command can reliably be killed because it does not depend on the child process successfully exiting to be terminated. However the whole reason I was using `pipe` is to provide debugging output for this feature. 

If we simply use `Timeout.timeout() { out = run }` and time out the call then we never get the output from the `run` command (because the execution fires before it can be set.

To meet our needs we need:

- To be able to reliably time out this command.
- To provide debugging output for when the command fails or times out.

The alternative that I came up with is to tell the `run` command to write to a file while it is running. If the command is killed via a timeout then the contents of the file are preserved and can be used for debugging. 
The updated timeout test fails on master, and passes on this branch.