Allow warnings to work with multiple classes
As written warnings are expected to be directly on a buildpack class. This is not always the case, for example the `RailsRunner` object can warn.

To fix this we need shared mutable state between everything that inherits from shell helpers. Since both instances and classes use these methods they should be shared between the two via a class variable.