Support TAP output for CI
The TAP specification says that there should be no whitespace in front of output. The `pipe` command auto indents output so that it looks correct with the build output.

This PR adds the ability to specify the object that is used for puts. We can pass in the `Kernel` object to force it to not indent output.

Related support ticket: 605130