[close #797] Do not carry asset pipeline fragment cache into runtime
When someone is using the asset pipeline, and they've disabled the config that allows them to build assets in production.

So when

```
config.assets.compile = false
```

Then the `tmp/cache/assets/sprockets/` folder is not needed at runtime and should be safe to delete, which will lower the overall slug size.

Do not merge until after Dreamforce.