Handle Rails 4 apps without sprockets-rails
When building a Rails 4 without `sprockets-rails`, asset compilation is
assumed to always be enabled, which would then cause an exception to be
raised when trying to determine whether the `sprockets` version is
vulnerable to CVE-2018-3760.

This commit ensures that Asset Compilation is properly detected as
being disabled when the `sprockets-rails` Gem is absent.

Closes #866.