Add support to Rails 6 by adding a new Language Pack
The current Rails 5 language pack work for Rails 6 applications but it
could not be used because of the version constraints. This was making
Rails 6 apps to be handled as generic rack apps and causing problems at
deploy if no SECRET_KEY_BASE was set.

By adding a new class we can update the constraint so new Rails 6 apps
can work as in the pre-releases.