Remove mutating test
The yarn_installer_spec mutates the current environment. It doesn't test anything other than that the yarn lib was added to the path, which doesn't even tell us if it's a valid test.

This test also had competing chdir calls which trigger a warning in Ruby (sometimes, but not always which is really surprising to me). This warning happens because the `LanguagePack::Ruby.new includes a `Dir.chdir` call in it.

Instead of working around this warning, we can safely get rid of this test as yarn is exercised directly in other integration tests.