Vendor libpq5.12.1 into Slugs
Libpq is the client library used to communicate to postgresql. There was a bug in libpq 5.11 that essentially said if there is a problem with a configuration value, for example if you tried to set `pool` (which should be an integer) to a string, then the value would be silently ignored. In libpq 5.12 this behavior was "fixed" so that if you try to set a value to an incorrect type it will raise an error:

```
PG::UnableToSend: no connection to the server
```

Or 

```
CONNECTION_BAD
```

Here is the commit that changed the behavior https://github.com/postgres/postgres/commit/e7a2217978d9cbb2149bfcb4ef1e45716cfcbefb

The largest failure mode is when someone accidentally forgets to put a space after their ERB tag in `database.yml` but before their comment. For example:


```
connect_timeout: <%= ENV['DB_CONNECT_TIMEOUT'] || 5 %># seconds
```

Would be evaluated to a string of "5# seconds". This will work in libpq5.11 but fail in libpq5.12. This can be fixed by adding a space between `%>` and `# seconds`.

This change was rolled out on our stack image, but then rolled back to customers getting failures in the middle of the night when they had seemingly changed nothing. 

By vendoring this library in the slug, we have more control around when the update will happen. Rather than breaking changes when the stack image rolls out which might happen in the middle of the night for some developers, this change will only be applied as customers are deploying their apps. If they have a release phase then the error will likely be caught there and prevented from going into production.

This behavior introduces a helpful warning message along with URLs for getting more information around this issue.

We do not anticipate that a majority of applications will be affected by this change, however those that are affected experience large problems. The intent of rolling this change out in a buildpack is to minimize the exposure of apps that will trigger this failing condition.

Eventually libpq5.12.1 will be the default version on the stack. This change works to prepare developers for that change.