Force user's bundler on the path in CI
Ruby 2.6+ ships with a version of bundler. If we do not ensure that the customer's `bundle` binary is on the path before the Ruby buildpack's then the buildpacks version may be used.

This PR prevents this problem by ensuring that the vendored customer version of bundler is always first in the path before any existing binaries and after the customers `bin/` dir.

# Do not modify or remove the line above.
# Everything below it will be ignored.
#
# Date:      Wed Mar 11 14:07:17 2020 -0500
#
# interactive rebase in progress; onto f20f977
# Last commands done (3 commands done):
#    pick 0b5e9f8 Force user's bundler on the path
#    squash d51fa3b Blerg
# No commands remaining.
# You are currently rebasing branch 'schneems/ci-fails-with-ruby-265' on 'f20f977'.
#
# Changes to be committed:
#       modified:   bin/support/ruby_test
#