Handle binary binstubs
Logic was added to check the "shebang" lines of binstubs. This was then reverted due to a bug #1019. This commit fixes this error by checking if a shebang line can be considered UTF-8 for purposes of comparison before we compare it. 

In the process I also refactored the code so that the logic around working with the binstub file is in BinstubWrapper. 

I added a test for the `BinstubWrapper#binary?` method as well as a higher level test that runs the whole check against the directory that contains the `which ruby` binary.