[changelog skip] ✂️ Hatchet#setup! (#1061)
Originally the Hatchet#setup! method only created an app and set config. Now it's responsible for also mutating files if there's a before_deploy action and for creating a git directory if using a non-git path.

Originally calling `setup!` was necessary before calling `deploy` if you wanted to do something like set config on an app before deploy. It was necessary because the Hatchet::Runner.new interface did not have a `config:` interface. That's no longer the case, so it's not needed.

Using this outside of the repo is dangerous. We should probably make it a private API. But...first lets remove it from our existing buildpacks.

While nothing in the ruby buildpack currently accidentally modifies the existing disk, it's purely by accident. If we were to copy and paste a test and modify it, it's possible that could change. Let's remove all use of `Hatchet#setup!` to be safe.