Move tarball logic to class for testing
Logic in the rake file is not tested. I want to start porting code over to custom classes that can be tested. Then we can use those classes with a minimal interface in the rakefile. It won't guarantee there are no regressions, but it will give much more confidence when modifying the logic.

In addition to moving this logic over to a custom class. The logic is changed so that it's tarballing contents from disk instead of from git remote so it can be fired from GitHub actions.