Add `rake buildpack:release` command
This command will run checks to ensure that the local copy is the same as the copy on GitHub. It then creates and pushes a tag to GitHub, then uses the heroku buildpack CLI to deploy the change.

Currently the process for deploying a buildpack involves manually making a tag, then calling the cli. This allows us to turn that into a one step process that we can add guards to (for instance checking changelog contents).

The majority of the logic for this rake task is in a custom class that is tested.