[close #1069] Remove Active Storage warnings
There are two active storage warnings:

- Warn when storage is set to :local
- Warn when ffmpeg is not present on the system

## Warn when storage is set to :local

The goal here was to prevent people from accidentally uploading assets only to find out that they disappear every 24 hours after the dyno resets.

While that's still a problem now, the default value is `:local` when you generate a new project. That means just because someone has this set to be local does not mean they're using active storage. That also means that this generates a useless warning message and clutters their deploy.

This PR removes that warning

## Warn when ffmpeg is not present on the system

If popper and ffmpeg are not on the system then Rails cannot generate previews for PDFs and videos which is a feature of active storage. To help, we added a warning if you're using active storage but don't have these dependencies.

Unfortunately, we can't tell if you're actually using these features or not, so there's no easy way to disable this warning if they are not needed.

In general these warnings cause a lot of clutter and confusion.