Use jvm buildpack to install JRuby dependencies (#1119)
* Use jvm buildpack to install JRuby dependencies

Currently, the jvm logic is duplicated in this buildpack and the jvm buildpack. By using the JVM buildpack we can reduce duplication and increase consistency across the Heroku platform.

* Use named variables instead of bash args

Previously the function compile_buildpack_v2 directly uses $1, $2, and $3. This change updates this function to instead use named local variables.

In addition, there was a typo in the output to the user.