Split bootstrap version from default version
Currently the bootstrapped version of Ruby is coupled to the default version we give customers who do not specify a version. 

With heroku-22 (based on Ubuntu 22.04) we cannot use the current default (3.0.3) because it cannot compile for heroku-22 https://bugs.ruby-lang.org/issues/18658.

This change provides a second mechanism for specifying the bootstrapped version.

GUS-W-10343881