Enable heroku-22 download presence check (#1290)
Adding this does two things. It turns on the download check for anyone trying to deploy to heroku-22. This will suggest a different stack if a binary is not found. For example if a customer wants to use 2.5.4 on heroku-22 it would suggest they try to deploy to heroku-18 or heroku-20 instead as heroku-22 doesn't have that binary.

In theory it could also affect customers on other stacks trying to use a version of Ruby not available on their stack (but present on heroku-22). If that happens they'll see heroku-22 in their stack suggestions. While it's technically a possibility the only supported stacks now are heroku-18 and heroku-20 and they have all the binaries (and more) that heroku-22 has.

GUS-W-10343881