Remove Ruby dependency on bin/release (#1308)
* Remove Ruby dependency on bin/release

On heroku-22 there is no longer a system Ruby so we cannot depend on it. We can either bootstrap as we do in `bin/compile` or use bash. Considering this logic is trivial to represent in bash that's what we will do.

* Fix test

Failed due to a change in the Ruby getting started guide. This is perhaps not a great test that needs to be re-visited. I believe I was just looking to ensure that `ruby -v` was working and wasn't system ruby.

* v241