## Default Ruby version for new apps is now 3.1.2
The [default Ruby version for new Ruby applications is 3.1.2](https://devcenter.heroku.com/articles/ruby-support#default-ruby-version-for-new-apps). You’ll only get the default if the application does not specify a ruby version.

Before we can make heroku-22 the default we need the defaults to work with heroku-22. Due to https://bugs.ruby-lang.org/issues/18658 Ruby 2.7 and 3.0 do not work on Heroku-22. As a result we are upgrading the default version of Ruby to 3.1.x.

This change will only impact customers with no Ruby version specified in their Gemfile.lock

GUS-W-11484844