Allow custom vendored Node URL
Follow-up to https://github.com/heroku/heroku-buildpack-ruby/pull/1406

Lets us ship custom builds, targeting different platforms than the
Heroku default.

This is slightly more complicated than the Ruby change, as the URL contains a lot of
platform-specific information, as well as interpolates the Node version.

My solution is to allow for a variable in the URL that gets substituted
with the version.

This allows for dynamic lookups just like normal, and avoids hardcoding to a single version.