Clean up release automation
The scripts in .github/scripts were for CNB releases which have been moved over to https://github.com/heroku/buildpacks-ruby and have not been used for some time.

The rake task to prepare releases has been moved to a GitHub action. The deploy task is still used, it's updated to work with the new standard CHANGELOG format.