Fix tests after bundler change
The Ruby buildpack uses the Ruby buildpack on CI. When the bundler versions changed in https://github.com/heroku/heroku-buildpack-ruby/commit/56ff138eea331620d3579370693dd39140a642cd it also changed the behavior of these tests (which use the local bundler version). That means that after the buildpack was deployed, it retroactively meant that tests on main were failing.

The issue with these tests is that they're using a feature of bundler that's been removed, specifically this change https://devcenter.heroku.com/changelog-items/2809. In these tests the Ruby version is specified in the Gemfile but not the Gemfile.lock:

- Gemfile https://github.com/sharpstone/mri_193/blob/master/Gemfile
- Gemfile.lock https://github.com/sharpstone/mri_193/blob/master/Gemfile.lock 

With the newer version of bundler, these tests now fail. To fix this tests will need to be updated or removed (based on their current relevance).