Let bundler >= 2.2 handle windows platform (#1469)
* Let bundler >= 2.2 handle windows platform

Fixes #1157

Because it has support for multiple platforms in Gemfile.lock. So deleting Gemfile.lock when one of the platforms is windows is not good.
Bundler's error message tells the user how to handle it.
```
Your bundle only supports platforms ["mingw"] but your local platform is x86_64-linux. Add the current platform to the lockfile with
`bundle lock --add-platform x86_64-linux` and try again.
```

* Test and document Windows bundler 2.2+ behavior

* Update changelogs/unreleased/windows_gemfile.md

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>

---------

Co-authored-by: Rune Philosof <57357936+runephilosof-abtion@users.noreply.github.com>
Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>