Stop bundling bootstrap Ruby (#1481)
Since this buildpack is written in Ruby, it needs a Ruby installation
to be installed in order to run. As such, during buildpack compile, a
bootstrap Ruby is downloaded and installed.

As a performance optimisation, this bootstrap Ruby was vendored into
the buildpack archive published to the Buildpack Registry, via the
`[[publish.Vendor]]` entries in `buildpack.toml`. Since Ruby installs
are OS-specific, this meant bundling three copies of Ruby, even though
only one would ever be used. With the old build system this was still a
net benefit, since the official language buildpacks would be downloaded
in advance when a build worker first booted.

However, with the new build system architecture, the build system
workers are no longer booted en-mass in advance, but instead started
on demand. This means there is no worker idle time, and so speculative
downloading of possibly-needed assets no longer makes sense.

As such, the bundling of multiple copies of Ruby in the buildpack
registry archive has now been removed.

This not only makes non-Ruby builds faster (since they have no need
for a Ruby installation at all), but will also speed up Ruby builds,
since they'll only need to download one bootstrap Ruby (for their
specific stack), rather than a version for every stack.

GUS-W-16184923.