Fix tests not executing (#1517)
* Emit log order

Issue #1505 spells out a problem where it appears ci-queue does not fail even when the test suite is broken. I'm unsure why this happens. To debug I'm emitting the tests that run so I can audit to make sure `rspec ./spec/helpers/yarn_installer_spec.rb:6` is executed.

If it's not, then I need to diagnose why. If it is, then perhaps there some ordering bug that's affecting the outcome.

* Update CI-queue

* Update redis version

* Use branch of ci-queue