Add a mechanism for documenting Ruby versions via automation
Our current deploy process looks like this for releasing a Ruby version:

- Manually make a PR to the Ruby buildpack with the version in the changelog
- Get it approved by another engineer
- PreRelease the buildpack via automation
- Merge the pre-release PR
- Pull changes locally and run deploy command `bundle exec rake buildpack:release`
- Then trigger S3 changes via running automations on https://github.com/heroku/docker-heroku-ruby-builder

In the future there's a desire to have the https://github.com/heroku/docker-heroku-ruby-builder generate an "inventory" file that tracks binary checksums and then the buildpack will require a PR to utilize new versions. However, that's a non-trivial amount of work and Ruby versions keep getting released on Fridays when I'm on vacation and no one else is around.

With this change the process can look like this:

- Trigger this new workflow which updates the CHANGELOG.md and runs the pre-release automation in one
- Merge that PR
- Pull changes locally and run deploy command `bundle exec rake buildpack:release`
- Then trigger S3 changes via running automations on https://github.com/heroku/docker-heroku-ruby-builder

Which is a substantial workflow savings. This task can be removed once we implement the full inventory workflow (sometime in the vague future), but this is a quality of life upgrade for the short term.