Ensure flat YAML build report (#1574)
The `bin/report` structure is flat. All keys must map to values that are un-nested. Ruby's `to_yaml` will happily serialize ruby objects:

```
irb(main):004> puts Gem::Version.new("3.4.2").to_yaml
--- !ruby/object:Gem::Version
version: 3.4.2
```

This commit fixes this problem by inspecting if the value being serialized is an object or not. If it is, it is then converted into a string.