Update default Ruby version to 3.3.8 (#1595)
* Update default Ruby version to 3.3.8

Also updates the "bootstrap" version of Ruby that the buildpack uses to run itself to match the default version. Previously, it was pinned to 3.1.x due to needing to support Bundler 1.x for the heroku-20.

* Fix error difference

```
       expected ArgumentError, got #<Encoding::CompatibilityError:"invalid byte sequence in UTF-8\n       Invalid string: ACENOCUMAROL,AZECAR,1 MG 30 COMPRIMIDOS, \xF1\xF1\xF3\xE9n\xF1\r\n"> with backtrace:
         # ./spec/helpers/shell_spec.rb:73:in `strip'
         # ./spec/helpers/shell_spec.rb:73:in `print'
```

We don't need to rely on the error class, as we are asserting against the message.