Report when version exists but our parser doesn't handle it
This is a layered approach, if the file has the RUBY VERSION but the match is empty, it will be recorded, we then try to extract the incorect line using a more permissive regex, but since we don't know why it's failing this might be wrong in subtly different ways. When the `count(failed_parse) == count(failed_contents)` we know we're permissive enough.