Capture when new and old Ruby versions are different
The new Ruby version infrastructure still does nothing, but in theory it should produce the same `version_for_download` as the old mechanism. This code tests that theory by capturing information.

The one known difference is that `bundle platform --ruby` will output a Ruby version if it's persisted in the Gemfile but not in the Gemfile.lock, this new code requires that the Ruby version ends up in the Gemfile.lock, part of this metrics process is determining how often people rely on that old behavior (where they set a Ruby version, but didn't persist it to their Gemfile.lock).