Refactor RubyVersion class (#1601)
* Refactor RubyVersion args

Prefer kwargs, remove open-ended options hash in favor of explicit kwarg.

* Inline initialize logic

* Remove non-public instance variables

These values do not need to be persisted as they're not used after initialization

* Remove unused variable

* Rename ivar and remove pub, method already exposed

* Refactor version increment logic

* Rename ivar to match method

* Remove unused ivar

* Derive version_for_download

* RubyVersion is now a simple store

The bundle parsing logic feeds into the RubyVersion class instead of being tightly coupled

* Use new interface

* Update method

The other method was removed, this is the new name

* Add test, passes on main, fails here

* Fix test, capture "pre" release version information

* Support pre-release versions without numbers

Convention says they have numbers, but it's not strictly required.

* Remove shell helper include

* Extract RubyVersion.default into a method

* Capture when Ruby version is the default