Add mocked testing for Cache class
Testing this has proved to be a pain. The internals rely on gnucp but I develop on a mac (bsdcp) which has different flags. Reproducing the copy behavior in Ruby is non-trivial and requires handrolling a lot without the benefit of a strong type system. 

This testing strategy verifies that we've instructed copying certain directories from/to with a specific behavior force: true/false. It uses mocks. This is heavily coupled but provides some hints about usage and expectations.