Encapsulate new_app? check
Currently this check relies on the metadata folder existance for whether or not it's considered a "new app." This relies on implicit order of when the Metadata class is initialized as this creates the folder.

By moving the check into the class, we have a stronger guarantee that the order of the check and the load is correct (though this also technically requires that the metadata is a singleton through the lifecycle of the buildpack.