Fix failing test regarding copy checking
The core problem was that if the directory we're coping into does not exist, we would create it using `mkpath` rather than copy it wholesale recursively. This means that the creation times are the times from `mkpath` instead of the copied directory. This is fixed by checking the target directory and either copying the internal files, or the whole directory.