Fix order of mtime setting
If mtime is set before the file is written, then the mtime is updated with the recent file creation time. This makes it so that deleting the mtime setting code in the test does not change the outcome of the test i.e. it's not testing what it looks like it's testing. This commit fixes that in two locations