Fix mtime test
In the event that we need to create a parent directory, we already know the file doesn't exist. If the directory also doesn't exist then it means that there are no conflicts to check for under it so we can wholesale copy the directory in its entirety. That means we should never need to call `.mkpath` so we'll never get the directories out of sync (permissions or file times).