Remove vendored buildpack detection
The old buildpack logic would pre-vendor a Ruby version at deploy in `vendor/ruby/$stack/$arch` of the buildpack dir. There were problems when the Ruby buildpack was specified to run twice so a check was added for that condition.

The logic was changed to always download the Ruby verison and not pre-vendor details in the PR https://github.com/heroku/heroku-buildpack-ruby/pull/1481.

Since that's no longer happening, we can avoid the check and just create a tempdir every time. This simplifies the logic.