Inline bash function
Fixing the previously hidden warning introduced a new error:

```
       remote: -----> Ruby app detected
       remote: /tmp/codon/tmp/buildpacks/ed3d993430eab40cdfcb6130d0909c19a1b9ae24/bin/compile: line 39: /tmp/tmp.Uu0SIOeXEb/bin/ruby: No such file or directory
```

What's happening here is that the `trap` to cleanup the bootstrap dir was firing on subshell exit, and since we were capturing the output of that function via `bootstrap_ruby=$(...)` it was creating the directory and immediately deleting it.

This was previously not a problem as the local variable scoping prevented the `rm -rf` from firing correctly. But fixing that variable scoping error introduced this problem. We can fix by moving that trap out of a sub-shell. Since this function is really only 3 lines long at this point, there's not much benefit in having it as a function, especially if it cannot provide the automatic cleanup behavior as desired.