Track unique version number
Both JRuby and MRI share the same `ruby.version` value. i.e. `3.4.2` to distinguish we are switching to the `version-for-download` which includes the information about the engine, this will make queries more clear whether they include jruby or not.

The prior value is moved to "spec" version as engines such as JRuby will implement to a spec such as `3.4.2`.

GUS-W-19310643