Remove error handling from printing
Comment https://github.com/heroku/heroku-buildpack-ruby/pull/1635/files#r2293616228

```
I'm not too sure about this error handling implementation.

If the env var isn't set, it's a buildpack bug, and with the :-'(unset) fallback removed that bug would correctly be caught by bash error on undefined env vars (which which file enables). As-is, this fallback masks the issue and causes bin/report to exit zero when it actually hit an internal error case.

Also, if the report file is missing, that's also an internal buildpack bug (and not a user facing error case).

Currently cytokine logs an error if a buildpack has a bin/report and it exits 1 (having set a minimal build report), which I think might be a better way to catch these cases? If we did still want to customise the report in this scenario, implementing it in cytokine (rather than each buildpack) would allow standardising across buildpacks more easily.
```

Adds in a check for when the exported env var HEROKU_RUBY_BUILD_REPORT_FILE is mutated to error.