Use verbose `git ls-files` exclude syntax
> "In the long form, the leading colon : is followed by an open parenthesis (, a comma-separated list of zero or more "magic words", and a close parentheses ), and the remainder is the pattern to match against the path."

It's not clear looking at the docs but the header itself qualifies as a "magic word"

```
exclude
After a path matches any non-exclude pathspec, it will be run through all exclude pathspecs (magic signature: ! or its synonym ^). If it matches, the path is ignored. When there is no non-exclude pathspec, the exclusion is applied to the result set as if invoked without any pathspec.
```

This allows us to use `(exclude)` as syntax here. https://github.com/heroku/heroku-buildpack-ruby/pull/1635#discussion_r2300592267